perf(desktop): switch test env from jsdom to happy-dom (-41% duration)
- Replace jsdom with happy-dom in vitest.config.ts (2-3x faster DOM sim)
- Guard Element polyfills in setup.ts for @vitest-environment node compat
- Extract buildDefaultOptions pure function tests to separate .test.ts
  file with @vitest-environment node (no DOM overhead)
- Replace remaining userEvent.click with fireEvent.click in
  SettingsPage and FormatOptionsPanel tests
- Remove jsdom dependency from package.json

Test suite: 281 tests, 19 files, 6.28s (was 10.57s with jsdom).
